Refactor QR Seed Provider to use dependency injection
Refactoring the QR Seed Provider so that it can be tested easier. And adding product flavors to the Android project. Now, mockDebug will always return the same QR code seed (and it won't even hit the server). The prodDebug variant should work as before.
